refactor: move struct functions to new structs.go
